aider: Add link to 'GlobePage' in 'IndexPage'.
# Aider chat conversation:

USER: make it so it's accessible from 'IndexPage'

ASSISTANT: To make the Globe accessible from the `IndexPage`, we need to add a link to the `GlobePage` in the `IndexPage`. We can do this by adding a `router-link` to the `GlobePage` in the `IndexPage` template.

Here is the *SEARCH/REPLACE* block:

src/pages/IndexPage.vue
```vue
<<<<<<< SEARCH
<template>
  <q-page class="row items-center justify-evenly">
    <example-component
      title="Example component"
      active
      :todos="todos"
      :meta="meta"
    ></example-component>
  </q-page>
</template>
=======
<template>
  <q-page class="row items-center justify-evenly">
    <example-component
      title="Example component"
      active
      :todos="todos"
      :meta="meta"
    ></example-component>
    <router-link to="/globe">Go to Globe Page</router-link>
  </q-page>
</template>
>>>>>>> REPLACE
```

This will add a link to the `GlobePage` on the `IndexPage`. When you click on the link, it will navigate to the `GlobePage`.